+++ stats/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@fastify/cors": "^11.0.0",
     "@fastify/url-data": "^6.0.3",
+    "@fastify/postgres": "^5.2.0",
     "@filecoin-station/spark-stats-db": "^1.0.0",
     "@sentry/node": "^9.5.0",
     "@sentry/profiling-node": "^9.5.0",
